flask-app-v4/requirements.txt
Flask==2.1.2
Werkzeug==2.0.3
prometheus_client==0.12.0
opentelemetry-api==1.5.0
opentelemetry-sdk
opentelemetry-instrumentation-wsgi
opentelemetry-exporter-jaeger==1.5.0
opentelemetry-instrumentation-flask

